requirements.txt
torch==1.13.1
diffusers==0.16.1
Pillow==9.4.0
gradio==3.22.0